adjisting
