+++ src/pages/staffList005/DialogButton.jsx
@@ -89,96 +89,96 @@
 // }
 
 import React from 'react';
-import { withStyles } from '@material-ui/core/styles';
 import Button from '@material-ui/core/Button';
-import Dialog from '@material-ui/core/Dialog';
-import MuiDialogTitle from '@material-ui/core/DialogTitle';
-import MuiDialogContent from '@material-ui/core/DialogContent';
-import MuiDialogActions from '@material-ui/core/DialogActions';
-import IconButton from '@material-ui/core/IconButton';
+import ButtonGroup from '@material-ui/core/ButtonGroup';
+import { makeStyles } from '@material-ui/core/styles';
 import CloseIcon from '@material-ui/icons/Close';
+import Paper from '@material-ui/core/Paper';
 import Typography from '@material-ui/core/Typography';
+import IconButton from '@material-ui/core/IconButton';
+import StaffList005_figure from './StaffList005_figure';
 
-const styles = (theme) => ({
+const styles = makeStyles((theme) => ({
   root: {
-    margin: 0,
-    padding: theme.spacing(2),
+    display: 'flex',
+    flexDirection: 'column',
+    alignItems: 'center',
+    '& > *': {
+      margin: theme.spacing(1),
+    },
   },
-  closeButton: {
-    position: 'absolute',
-    right: theme.spacing(1),
-    top: theme.spacing(1),
-    color: theme.palette.grey[500],
+  margin: {
+    bottom: theme.spacing(16),
+    left: theme.spacing(5),
+    background:'#000000',
+    float: 'right',
   },
-});
+  icon: {
+    color: '#ffffff',
+  },
+  Button: {
+    background:'rgb(120,144,156)',
+    borderRadius: '0px 0px 0px 0px',
+  },
+  Button2: {
+    background:'rgb(120,144,156)',
+    left: 45,
+    color: '#ffffff',
+    borderRadius: '0px 0px 0px 0px',
+    float: 'right',
+  },
+  Button3: {
+    background:'rgb(120,144,156)',
+    left: 45,
+    color: '#ffffff',
+    borderRadius: '0px 0px 0px 0px',
+    float: 'right',
+  },
+  Button4: {
+    top: theme.spacing(2),
+    right: theme.spacing(6),
+    color: '#000000',
+    height: '30px',
+  },
+  ButtonGroup: {
+    margin: '10px 200px 10px 0px',
+    borderRadius: '0px 0px 0px 0px',
+  },
+  Paper: {
+    padding: theme.spacing(7),
+    margin: '16px',
+    width: '95%',
+    borderRadius: '12px 12px 12px 12px',
+  },
+  title: {
+    margin: '10px 0px 10px 0px',
+  },
+}));
 
-const DialogTitle = withStyles(styles)((props) => {
-  const { children, classes, onClose, ...other } = props;
+// export default function BasicButtonGroup() {
+const DialogButton = () => {
+  const classes = styles();
+
   return (
-    <MuiDialogTitle disableTypography className={classes.root} {...other}>
-      <Typography variant="h6">{children}</Typography>
-      {onClose ? (
-        <IconButton aria-label="close" className={classes.closeButton} onClick={onClose}>
-          <CloseIcon />
-        </IconButton>
-      ) : null}
-    </MuiDialogTitle>
-  );
-});
-
-const DialogContent = withStyles((theme) => ({
-  root: {
-    padding: theme.spacing(2),
-  },
-}))(MuiDialogContent);
-
-const DialogActions = withStyles((theme) => ({
-  root: {
-    margin: 0,
-    padding: theme.spacing(1),
-  },
-}))(MuiDialogActions);
-
-export default function CustomizedDialogs() {
-  const [open, setOpen] = React.useState(false);
-
-  const handleClickOpen = () => {
-    setOpen(true);
-  };
-  const handleClose = () => {
-    setOpen(false);
-  };
-
-  return (
-    <div>
-      <Button variant="outlined" color="primary" onClick={handleClickOpen}>
-        Open dialog
-      </Button>
-      <Dialog onClose={handleClose} aria-labelledby="customized-dialog-title" open={open}>
-        <DialogTitle id="customized-dialog-title" onClose={handleClose}>
-          Modal title
-        </DialogTitle>
-        <DialogContent dividers>
-          <Typography gutterBottom>
-            Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis
-            in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros.
-          </Typography>
-          <Typography gutterBottom>
-            Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Vivamus sagittis
-            lacus vel augue laoreet rutrum faucibus dolor auctor.
-          </Typography>
-          <Typography gutterBottom>
-            Aenean lacinia bibendum nulla sed consectetur. Praesent commodo cursus magna, vel
-            scelerisque nisl consectetur et. Donec sed odio dui. Donec ullamcorper nulla non metus
-            auctor fringilla.
-          </Typography>
-        </DialogContent>
-        <DialogActions>
-          <Button autoFocus onClick={handleClose} color="primary">
-            Save changes
-          </Button>
-        </DialogActions>
-      </Dialog>
+    <div className={classes.root}>
+      {/* <Paper variant="outlined" className={classes.Paper}> */}
+        {/* <Typography variant="h2" className={classes.title}>スタッフ詳細</Typography> */}
+        <Typography variant="h5">テスト太郎</Typography>
+        {/* <IconButton aria-label="delete" className={classes.margin}>
+          <CloseIcon className={classes.icon} />
+        </IconButton> */}
+        <div>
+          <ButtonGroup className={classes.ButtonGroup}　size="small" variant="contained" aria-label="contained primary button group">
+            <Button className={classes.Button}>スキル</Button>
+            <Button className={classes.Button}>経歴</Button>
+            <Button className={classes.Button}>キャリアパス</Button>
+            <Button className={classes.Button}>スタッフ情報</Button>
+          </ButtonGroup>
+          <Button variant="contained"　className={classes.Button2}>スキルシート出力</Button>
+          <Button variant="contained"　className={classes.Button3}>引合登録</Button>
+        </div>
+      {/* </Paper> */}
     </div>
   );
 }
+    export default DialogButton;
